Dmirano/3778 apd overview components restructure (#3896)
* layout restructure

* updated tests

* dashboard refactor

* update tests

* added login refactor

* updated tests

* added admin refactor

* updated tests

* route refactor

* route refactor to /pages/

* updated tests (not finishes yet)

* apd components refactor

* updated tests (route still causing issues)

* updated tests

* update tests, all pass now

* fixed merge conflicts

* relocated apd overview components

* updated tests and renamed to ApdOverview
