+++ web/src/containers/ApdApplication.js
@@ -13,7 +13,7 @@ import Sidebar from '../layout/nav/Sidebar';
 import UnexpectedError from './UnexpectedError';
 import { setApdToSelectOnLoad, selectApd } from '../actions/app';
 
-import ApdPageRoutes from './ApdPageRoutes';
+import ApdPageRoutes from '../pages/apd/ApdPageRoutes';
 import Loading from '../components/Loading';
 
 import { getAPDId } from '../reducers/apd';
