+++ web/src/pages/apd/ApdPageRoutes.js
@@ -7,17 +7,17 @@ import {
   useRouteMatch as actualUseRouteMatch
 } from 'react-router-dom';
 
-import ApdHeader from '../layout/header/ApdHeader';
-import Activities from './activity/All';
-import EntryPage from './activity/EntryPage';
-import AssurancesAndCompliance from './AssurancesAndCompliance';
-import Export from './ApdExport';
-import ApdSummary from './ApdSummary';
-import ExecutiveSummary from './ExecutiveSummary';
-import PreviousActivities from './PreviousActivities';
-import ProposedBudget from './ProposedBudget';
-import ScheduleSummary from './ScheduleSummary';
-import StateProfile from '../components/ApdStateProfile';
+import ApdHeader from '../../layout/header/ApdHeader';
+import Activities from '../../containers/activity/All';
+import EntryPage from '../../containers/activity/EntryPage';
+import AssurancesAndCompliance from '../../containers/AssurancesAndCompliance';
+import Export from '../../containers/ApdExport';
+import ApdOverview from './apd-overview/ApdOverview';
+import ExecutiveSummary from '../../containers/ExecutiveSummary';
+import PreviousActivities from '../../containers/PreviousActivities';
+import ProposedBudget from '../../containers/ProposedBudget';
+import ScheduleSummary from '../../containers/ScheduleSummary';
+import StateProfile from '../../components/ApdStateProfile';
 
 const ApdPageRoutes = ({ apdId, useRouteMatch }) => {
   const { path } = useRouteMatch();
@@ -40,7 +40,7 @@ const ApdPageRoutes = ({ apdId, useRouteMatch }) => {
         </Route>
 
         <Route path={`${path}/apd-overview`}>
-          <ApdSummary />
+          <ApdOverview />
         </Route>
 
         <Route path={`${path}/previous-activities`}>
