+++ web/src/pages/apd/apd-overview/ApdOverview.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React, { useState } from 'react';
 import { Alert, ChoiceList, TextField } from '@cmsgov/design-system';
 import { connect } from 'react-redux';
-import DeleteModal from '../components/DeleteModal';
+import DeleteModal from '../../../components/DeleteModal';
 
 import {
   addYear,
@@ -12,14 +12,14 @@ import {
   setNarrativeForHIT,
   setNarrativeForMMIS,
   setProgramOverview
-} from '../actions/editApd';
-import RichText from '../components/RichText';
-import Instruction from '../components/Instruction';
-import { Section } from '../components/Section';
-import { t } from '../i18n';
-import { selectSummary } from '../reducers/apd.selectors';
+} from '../../../actions/editApd';
+import RichText from '../../../components/RichText';
+import Instruction from '../../../components/Instruction';
+import { Section } from '../../../components/Section';
+import { t } from '../../../i18n';
+import { selectSummary } from '../../../reducers/apd.selectors';
 
-const ApdSummary = ({
+const ApdOverview = ({
   addApdYear,
   name,
   narrativeHIE,
@@ -173,7 +173,7 @@ const ApdSummary = ({
   );
 };
 
-ApdSummary.propTypes = {
+ApdOverview.propTypes = {
   addApdYear: PropTypes.func.isRequired,
   removeApdYear: PropTypes.func.isRequired,
   name: PropTypes.string.isRequired,
@@ -202,6 +202,6 @@ const mapDispatchToProps = {
   setOverview: setProgramOverview
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(ApdSummary);
+export default connect(mapStateToProps, mapDispatchToProps)(ApdOverview);
 
-export { ApdSummary as plain, mapStateToProps, mapDispatchToProps };
+export { ApdOverview as plain, mapStateToProps, mapDispatchToProps };
