+++ web/src/pages/routes.js
@@ -1,7 +1,7 @@
 import NoMatch from '../components/NoMatch';
 import Dashboard from '../containers/Dashboard';
 import ApdApplication from '../containers/ApdApplication';
-import ApdViewOnly from '../containers/viewOnly/Apd';
+import ApdViewOnly from './apd/ApdReadOnly';
 import LoginApplication from '../containers/LoginApplication';
 import ManageAccount from './admin/state-admin/ManageAccount';
 import StateAdmin from './admin/state-admin/StateAdmin';
